add and update test cases for api_version support
