+++ test/sanity-check/sanity.js
@@ -15,16 +15,16 @@ require('./api/contentType-test')
 require('./api/asset-test')
 require('./api/extension-test')
 require('./api/entry-test')
+require('./api/variantGroup-test')
+require('./api/variants-test')
+require('./api/ungroupedVariants-test')
+require('./api/entryVariants-test')
 require('./api/bulkOperation-test')
 require('./api/webhook-test')
 require('./api/workflow-test')
 require('./api/globalfield-test')
 require('./api/release-test')
 require('./api/label-test')
-require('./api/variantGroup-test')
-require('./api/variants-test')
-require('./api/ungroupedVariants-test')
-require('./api/entryVariants-test')
 require('./api/contentType-delete-test')
 require('./api/delete-test')
 require('./api/team-test')
